Refactor refresh token endpoing and cleanup
