+++ src/main/kotlin/com/github/jaksa97/LeafSaver_Kotlin/models/auth/RefreshTokenRequest.kt
@@ -1,0 +1,5 @@
+package com.github.jaksa97.LeafSaver_Kotlin.models.auth
+
+data class RefreshTokenRequest(
+    val refreshToken: String
+)
