+++ src/main/kotlin/com/github/jaksa97/LeafSaver_Kotlin/models/dtos/producer/ProducerDto.kt
@@ -1,8 +1,8 @@
 package com.github.jaksa97.LeafSaver_Kotlin.models.dtos.producer
 
 import io.swagger.v3.oas.annotations.media.Schema
-import lombok.Value
 import lombok.Builder
+import lombok.Value
 
 @Value
 @Builder
