+++ src/main/kotlin/com/github/jaksa97/LeafSaver_Kotlin/models/mappers/UserMapper.kt
@@ -18,7 +18,5 @@ interface UserMapper {
         )
     }
 
-    //fun toEntity(userDto: UserDto): UserEntity
-
     fun toEntity(userSaveDto: UserSaveDto): UserEntity
 }
