+++ src/main/kotlin/com/github/jaksa97/LeafSaver_Kotlin/repositories/CureRepository.kt
@@ -3,11 +3,9 @@ package com.github.jaksa97.LeafSaver_Kotlin.repositories
 import com.github.jaksa97.LeafSaver_Kotlin.models.entities.CureEntity
 import com.github.jaksa97.LeafSaver_Kotlin.models.entities.DiseaseEntity
 import com.github.jaksa97.LeafSaver_Kotlin.models.entities.DrugEntity
-import jakarta.persistence.Id
 import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 
-interface CureRepository: JpaRepository<CureEntity, Int>, JpaSpecificationExecutor<CureEntity> {
+interface CureRepository: JpaRepository<CureEntity, Int> {
 
     override fun findAll(): List<CureEntity>
