+++ src/main/kotlin/com/github/jaksa97/LeafSaver_Kotlin/repositories/DiseaseRepository.kt
@@ -5,11 +5,10 @@ import com.github.jaksa97.LeafSaver_Kotlin.utils.Queries
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.data.jpa.repository.Query
 import java.util.*
 
-interface DiseaseRepository: JpaRepository<DiseaseEntity, Int>, JpaSpecificationExecutor<DiseaseEntity> {
+interface DiseaseRepository: JpaRepository<DiseaseEntity, Int> {
 
     @Query(Queries.DisaseNameSarchQuery)
     fun findAll(name: String?, pageable: Pageable): Page<DiseaseEntity>
