+++ src/main/kotlin/com/github/jaksa97/LeafSaver_Kotlin/repositories/DrugRepository.kt
@@ -6,11 +6,10 @@ import com.github.jaksa97.LeafSaver_Kotlin.utils.Queries
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.data.jpa.repository.Query
 import java.util.*
 
-interface DrugRepository: JpaRepository<DrugEntity, Int>, JpaSpecificationExecutor<DrugEntity> {
+interface DrugRepository: JpaRepository<DrugEntity, Int> {
 
     @Query(Queries.DrugNameSearchQuery)
     fun findAll(name: String?, pageable: Pageable): Page<DrugEntity>
